+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/comments/styles.tsx
@@ -5,7 +5,9 @@ export const DocsCommentsStyle = createGlobalStyle<{
   currentUserAvatarUrl?: string;
 }>`
   .--docs--main-editor.bn-root,
-  .--docs--main-editor.bn-root .ProseMirror {
+  .--docs--main-editor.bn-root .ProseMirror,
+  .--docs--comments-sidebar.bn-root,
+  .--docs--comments-sidebar.bn-root .ProseMirror {
     // Comments marks in the editor
     .bn-editor {
       // Resets blocknote comments styles
@@ -38,6 +40,22 @@ export const DocsCommentsStyle = createGlobalStyle<{
               );
             }
           }
+
+          .bn-thread-mark[data-orphan='true']:has(> .bn-thread-mark-selected) {
+            background-color: color-mix(
+              in srgb,
+              var(--c--contextuals--background--palette--blue-1--tertiary) 40%,
+              transparent
+            );
+            border-bottom: 2px solid
+              var(--c--contextuals--background--palette--blue-1--secondary);
+
+            mix-blend-mode: multiply;
+
+            transition:
+              background-color var(--c--globals--transitions--duration),
+              border-bottom-color var(--c--globals--transitions--duration);
+          }
         `}
 
       [data-show-selection] {
@@ -56,7 +74,7 @@ export const DocsCommentsStyle = createGlobalStyle<{
       min-width: calc(min(400px, 90vw));
       max-width: calc(min(400px, 90vw));
       max-height: calc(min(500px, 60vh));
-      padding: 8px;
+      padding: var(--c--globals--spacings--xxs) var(--c--globals--spacings--xxxs);
       box-shadow: 0px 6px 18px 0px #00001229;
       margin-left: 20px;
       margin-right: 20px;
@@ -81,10 +99,20 @@ export const DocsCommentsStyle = createGlobalStyle<{
         display: none;
       }
 
+      & .bn-thread-comments {
+        gap: var(--c--globals--spacings--xxxs);
+      }
+
       .bn-thread-comment {
         padding: 8px;
         flex-wrap: nowrap;
         gap: 0px;
+        flex-direction: column;
+        align-items: initial;
+
+        & > div:first-child {
+          flex-direction: row;
+        }
 
         & .bn-editor {
           padding-left: var(--c--globals--spacings--lg);
@@ -100,7 +128,7 @@ export const DocsCommentsStyle = createGlobalStyle<{
             padding: var(--c--globals--spacings--0)
               var(--c--globals--spacings--st);
             background: none;
-            border: 1px solid var(--c--globals--colors--gray-300);
+            border: 1px solid var(--c--contextuals--border--semantic--neutral--tertiary);
             border-radius: var(--c--globals--spacings--st);
             height: var(--c--globals--spacings--md);
           }
@@ -131,6 +159,8 @@ export const DocsCommentsStyle = createGlobalStyle<{
 
           // Date
           span.mantine-focus-auto {
+            font-weight: 400;
+            margin-left: var(--c--globals--spacings--2xs) !important;
           }
 
           .bn-comment-actions {
@@ -139,6 +169,9 @@ export const DocsCommentsStyle = createGlobalStyle<{
 
             .mantine-Button-root {
               background-color: transparent;
+              height: var(--c--globals--spacings--md);
+              width: var(--c--globals--spacings--md);
+              padding: var(--c--globals--spacings--0);
 
               &:hover {
                 background-color: var(--c--globals--colors--gray-100);
@@ -168,7 +201,7 @@ export const DocsCommentsStyle = createGlobalStyle<{
               height: var(--c--globals--spacings--md);
               padding-inline: var(--c--globals--spacings--st);
 
-              &[data-test='save'] {
+              &:first-child {
                 border: 1px solid
                   var(--c--contextuals--background--semantic--brand--primary);
                 background: var(
@@ -179,7 +212,7 @@ export const DocsCommentsStyle = createGlobalStyle<{
                 );
               }
 
-              &[data-test='cancel'] {
+              &:last-child {
                 background: white;
                 border: 1px solid
                   var(--c--contextuals--border--surface--primary);
@@ -214,6 +247,11 @@ export const DocsCommentsStyle = createGlobalStyle<{
           background-repeat: no-repeat;
           background-size: cover;
         }
+
+        & .bn-block-content:has(.ProseMirror-trailingBreak:only-child):after {
+          color: var(--c--contextuals--content--semantic--neutral--tertiary);
+          font-style: normal;
+        }
       }
 
       // Actions button send comment
@@ -223,6 +261,7 @@ export const DocsCommentsStyle = createGlobalStyle<{
 
         .bn-action-toolbar.bn-comment-actions {
           border: none;
+          background-color: transparent;
 
           button {
             font-size: 0;
@@ -256,6 +295,75 @@ export const DocsCommentsStyle = createGlobalStyle<{
 
           .ProseMirror.bn-editor {
             cursor: text;
+          }
+        }
+      }
+    }
+  }
+
+  /**
+  * Styles for the comments sidebar
+  */
+  .--docs--comments-sidebar.bn-root{
+    flex: 1;
+    min-height: 0;
+    overflow: auto;
+
+    .bn-threads-sidebar {
+      gap: 0;
+      border-radius: 0;
+
+      .bn-thread-expand-prompt p {
+        font-size: var(--c--globals--font--sizes--xs);
+      }
+
+      .bn-thread {
+        margin: 0;
+        max-width: 100%;
+        width: 100%;
+        min-width: 0;
+        overflow: visible;
+        padding: var(--c--globals--spacings--xxs) var(--c--globals--spacings--xxxs);
+        border: none;
+        border-radius: 0;
+        border-bottom: 1px solid var(--c--contextuals--border--surface--primary);
+
+        &.selected {
+          border: none;
+          background: var(--c--contextuals--background--semantic--neutral--tertiary);
+          max-height: none;
+        }
+
+        &:hover {
+          background: var(--c--contextuals--background--semantic--neutral--tertiary);
+        }
+
+        & .bn-header-text {
+          display: none;
+        }
+
+        &.bn-thread-orphaned {
+          & .bn-header-text {
+            display: block;
+            padding-inline: var(--c--globals--spacings--xs);
+          }
+        }
+
+        .bn-thread-comment {
+          padding: var(--c--globals--spacings--xs);
+
+          &:has(.bn-comment-actions) {
+            & > .mantine-Group-root:first-child {
+              background: linear-gradient(
+                to left,
+                var(--c--contextuals--background--semantic--neutral--tertiary) 90%,
+                rgba(255, 255, 255, 0) 100%
+              );
+            }
+
+            .bn-menu-dropdown {
+              box-shadow: 0px 0px 6px 0px #0000911a;
+            }
           }
         }
       }